refactor: change frontend structure
