+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import { RouterProvider } from 'react-router-dom'
 import { ChakraProvider } from '@chakra-ui/react'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { ColorModeProvider } from '@/components/ui/color-mode'
+import { ColorModeProvider } from '@/components/ui/ColorMode'
 import { AuthProvider } from '@/contexts/AuthContext'
 import { router } from '@/router'
 import { system } from '@/theme'
