+++ frontend/src/components/ProtectedRoute/index.tsx
