+++ frontend/src/components/TransactionList/transaction-list.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import { ChakraProvider, defaultSystem } from '@chakra-ui/react'
-import { TransactionList } from './transaction-list'
+import { TransactionList } from '.'
 import type { TransactionResponse } from '@/types'
 
 function renderWithProviders(ui: React.ReactElement) {
