+++ frontend/src/i18n/locales/pt.json
@@ -8,11 +8,11 @@
     "save": "Salvar",
     "delete": "Excluir",
     "close": "Fechar",
-    "skipToContent": "Pular para o conteudo principal"
+    "skipToContent": "Pular para o conteúdo principal"
   },
   "nav": {
     "dashboard": "Painel",
-    "transactions": "Transacoes"
+    "transactions": "Transações"
   },
   "auth": {
     "signIn": "Entrar",
@@ -25,63 +25,72 @@
     "emailPlaceholder": "Digite seu e-mail",
     "password": "Senha",
     "passwordPlaceholder": "Digite sua senha",
-    "passwordHint": "Minimo de 8 caracteres",
+    "passwordHint": "Mínimo de 8 caracteres",
     "firstName": "Nome",
-    "firstNamePlaceholder": "Joao",
+    "firstNamePlaceholder": "João",
     "lastName": "Sobrenome",
     "lastNamePlaceholder": "Silva",
-    "noAccount": "Nao tem uma conta?",
+    "noAccount": "Não tem uma conta?",
     "createOne": "Criar uma",
-    "haveAccount": "Ja tem uma conta?",
+    "haveAccount": "Já tem uma conta?",
     "signInLink": "Entrar",
     "loginFailed": "Falha no login. Por favor, tente novamente.",
     "registerFailed": "Falha no cadastro. Por favor, tente novamente."
   },
   "header": {
     "appName": "FinTech Wallet",
-    "greeting": "Ola, {{name}}",
+    "greeting": "Olá, {{name}}",
     "tagline": "Gerencie sua carteira digital com facilidade"
   },
   "dashboard": {
     "welcomeBack": "Bem-vindo de volta, {{name}}!",
-    "financialOverview": "Aqui esta seu resumo financeiro",
+    "financialOverview": "Aqui está seu resumo financeiro",
     "loadingBalance": "Carregando seu saldo...",
-    "quickActions": "Acoes Rapidas",
-    "addCredit": "+ Adicionar Credito",
-    "addDebit": "- Adicionar Debito",
-    "viewAllTransactions": "Ver Todas as Transacoes",
+    "quickActions": "Ações Rápidas",
+    "addCredit": "+ Adicionar Crédito",
+    "addDebit": "- Adicionar Débito",
+    "viewAllTransactions": "Ver Todas as Transações",
     "currentBalance": "Saldo Atual"
   },
   "transactions": {
-    "title": "Transacoes",
-    "subtitle": "Visualize e gerencie suas transacoes",
-    "transactionHistory": "Historico de Transacoes",
-    "loadingTransactions": "Carregando transacoes...",
-    "newTransaction": "Nova Transacao",
-    "quickTransaction": "Transacao Rapida",
-    "credit": "Credito",
-    "debit": "Debito",
-    "creditButton": "Credito (+)",
-    "debitButton": "Debito (-)",
+    "title": "Transações",
+    "subtitle": "Visualize e gerencie suas transações",
+    "transactionHistory": "Histórico de Transações",
+    "loadingTransactions": "Carregando transações...",
+    "newTransaction": "Nova Transação",
+    "quickTransaction": "Transação Rápida",
+    "credit": "Crédito",
+    "debit": "Débito",
+    "creditButton": "Crédito (+)",
+    "debitButton": "Débito (-)",
     "type": "Tipo",
     "amount": "Valor (R$)",
     "amountPlaceholder": "Digite o valor",
-    "createTransaction": "Criar Transacao",
+    "createTransaction": "Criar Transação",
     "creatingTransaction": "Criando...",
-    "createSuccess": "Transacao criada com sucesso!",
-    "createFailed": "Falha ao criar transacao",
+    "createSuccess": "Transação criada com sucesso!",
+    "createFailed": "Falha ao criar transação",
     "filterAll": "Todas",
-    "filterCredit": "Credito",
-    "filterDebit": "Debito",
-    "noTransactions": "Nenhuma transacao ainda",
-    "noTransactionsDescription": "Crie sua primeira transacao para comecar a acompanhar suas financas."
+    "filterCredit": "Crédito",
+    "filterDebit": "Débito",
+    "noTransactions": "Nenhuma transação ainda",
+    "noTransactionsDescription": "Crie sua primeira transação para começar a acompanhar suas finanças."
   },
   "accessibility": {
     "toggleTheme": "Alternar modo de cor",
     "selectLanguage": "Selecionar idioma",
     "closeModal": "Fechar modal",
-    "transactionType": "Tipo de transacao",
-    "filterTransactions": "Filtrar transacoes",
-    "mainNavigation": "Navegacao principal"
+    "transactionType": "Tipo de transação",
+    "filterTransactions": "Filtrar transações",
+    "mainNavigation": "Navegação principal"
+  },
+  "errors": {
+    "invalidCredentials": "Credenciais inválidas",
+    "emailAlreadyExists": "E-mail já cadastrado",
+    "serverError": "Erro do servidor",
+    "balanceFetchFailed": "Falha ao buscar saldo",
+    "networkError": "Erro de conexão",
+    "unauthorized": "Não autorizado",
+    "sessionExpired": "Sessão expirada"
   }
 }
